Pin the FP32 state contract of the fused delta-rule paths (#464)
## Human Note

## Agent note

Both reported failures reproduce only with FP16 Q/K/V and are the two directions of one policy: every
fused GDN/KDA path (Triton, CuTe, and Mega) carries the recurrent state and its cotangent in FP32 with
FP32 decay, but stages the chunk-entry state and the per-chunk state-cotangent tape in the Q/K/V dtype
because they feed tensor-core GEMMs. An FP32 state element of 65536 is carried correctly (1200.33 after
the decay) yet becomes inf when the unscaled chunk-entry state is read as an FP16 operand, and 0 * inf
poisons the chunk; a 2^-25 final-state cotangent flushes to zero in the FP16 tape before the gate, dK,
and dV state terms. There is no decay-first formulation for the W*h and Q*h GEMMs (exp(g_t - g_end) >= 1
would overflow instead), and a wider tape cannot feed the dK/dV GEMMs without a conversion pass, so the
contract is documented rather than paid for with a broad kernel cost. BF16 shares the FP32 exponent
range and passes the exact dynamic-range regressions on both backends and both families.

The one gratuitous narrowing found while mapping the dataflow is in the portable Triton GDN WY kernel,
which multiplied and reduced the <h, dh> gate term in the source dtype; the KDA Triton and CuTe kernels
already contract it in FP32. With zero Q/K/V and positive state/cotangent, the isolated state term of
dgate improves from 0.02--0.34 eps to 0.001--0.009 eps, and with cancellation from 4.5--4.9 to
0.15--2.3 BF16 eps, matching the CuTe path exactly.

The Mega KDA backend is the one genuine deviation from the FP32 carry: it applies the per-chunk decay
through a b16 diagonal MMA (state and dstate), re-rounding the carried state every 16 tokens. Model-range
impact is modest (final-state rel-RMS 3.6e-3 vs 2.7e-3 fused at T=256 BF16), so it is documented and
pinned with a strict xfail instead of re-plumbing the kernel here.

## Performance

Portable GDN backward, T=8192, H=8, BF16, GB200, triton do_bench (5 samples, 2 valid interleaved A/B
rounds): 796--798 us with the FP32 contraction vs 795--797 us before; no measurable cost.

## Test Plan

```bash
gpu-run --timeout 300 auto -- pytest -n 6 -q test/linear/test_gdn_chunk_fused.py test/linear/test_delta_rule_state_precision.py test/gdn/mega test/kda/mega/test_delta_rule_numerics.py -p no:cacheprovider -W ignore
gpu-run --timeout 300 auto -- python agent_space/repro_state_precision.py
gpu-run --timeout 300 auto -- python agent_space/measure_dg_last.py positive
```

+++ attn_gym/linear/gdn/bwd/triton/chunk_gdn_bwd_wy.py
@@ -152,7 +152,8 @@ def chunk_gdn_bwd_dqkwg_kernel(
             other=0.0,
         )
 
-        b_dg_last += tl.sum(b_h.to(b_dh.dtype) * b_dh)
+        # Accumulate the state term of dGate in FP32 rather than the source dtype.
+        b_dg_last += tl.sum(b_h.to(tl.float32) * b_dh.to(tl.float32))
         b_ds += tl.dot(b_do, tl.trans(b_v_new))
         b_dq += tl.dot(b_do, b_h)
         b_dk += tl.dot(b_v_new.to(b_dh.dtype), b_dh)
